code/atomic-tomorrow-creator/scripts/requirements.txt
# Portrait slicing script dependencies
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0